[13.0][FIX] ddmrp: do not consider internal reservations.
Operations done from locations within the buffer area/location
should not affect the on hand position of the buffer.

With this change, we rely on `stock.move.line`'s to compute the
qty reserved instead of using the `reserved_quantity` on the quant
which is the most accurate way to get the reserved qty. This also
eliminates `stock_available_unreserved` dependency.