Disable unarchiving buffer on product activation
Having the stock buffer automatically unarchived when the state of a
product changes is dangerous because it can lead to product being
ordered again when it should not anymore.

The issue encountered seemed to be related to how Odoo handles product
variant. And some changes in the product attribute line of a product
would call a write on the active field of related products variant.

An improvement that could be added (maybe coming soon) is to have on
the product form the information of existing archived stock buffers.